added some styling to Dashboard and added route to Reservation Form
