+++ package.json
@@ -56,6 +56,7 @@
     "jsonwebtoken": "^9.0.2",
     "passport": "^0.6.0",
     "passport-google-oidc": "^0.1.0",
+    "passport-local": "^1.0.0",
     "pg": "^8.11.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
