+++ src/App.jsx
@@ -4,6 +4,7 @@ import Dashboard from "./components/Dashboard.jsx";
 import Signup from "./components/Signup.jsx";
 import Login from "./components/Login.jsx";
 import Profile from "./components/Profile.jsx";
+import ReservationForm from "./components/ReservationForm.jsx"
 
 const App = () => {
   return (
@@ -14,6 +15,7 @@ const App = () => {
           <Route path="/signup" element={<Signup />} />
           <Route path="/dashboard" element={<Dashboard />} />
           <Route path="/profile" element={<Profile />} />
+          <Route path="/reservation" element={<ReservationForm/>}/>
         </Routes>
       </BrowserRouter>
     </div>
