+++ src/components/Navbar.css
@@ -1,9 +1,23 @@
 .Navbar{
     display:flex;
     justify-content:space-around;
+    margin:0;
     align-items:center;
     width:100%;
     padding:5px;
-    background-color:grey;
+    background-color:rgb(134, 83, 83);
     color:white;
+    font-size:20px;
+
+    button{
+        border:none;
+        font-size:20px;
+        color:rgb(222, 189, 189);
+        background-color:rgb(134, 83, 83);
+    }
 }
+
+body {
+    margin: 0;
+    padding: 0;
+  }
