+++ src/components/ReservationForm.jsx
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import Navbar from './Navbar.jsx'
+const ReservationForm = ()=>{
+
+    return(
+        <div>
+            <Navbar/>
+        </div>
+    )
+}
+
+export default ReservationForm;
